Move vs/game to template
Not VS specific anymore
