+++ template/CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_minimum_required(VERSION 3.1)
+set(32BLIT_PATH "../" CACHE PATH "Path to 32blit.cmake")
+project (game)
+include (${32BLIT_PATH}/32blit.cmake)
+blit_executable (game game.cpp game.hpp)
