+++ template/game.hpp
@@ -1,0 +1,1 @@
+#include "32blit.hpp"
