--- vs/game.sln
@@ -1,58 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29613.14
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "32blit", "32blit\32blit.vcxproj", "{280015A3-83D7-416F-A181-7CF88BCCE12F}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "32blit-sdl", "32blit-sdl\32blit-sdl.vcxproj", "{85DB89B5-DEA0-4429-B41E-0CAA7DB907BA}"
-	ProjectSection(ProjectDependencies) = postProject
-		{280015A3-83D7-416F-A181-7CF88BCCE12F} = {280015A3-83D7-416F-A181-7CF88BCCE12F}
-	EndProjectSection
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "game", "game\game.vcxproj", "{3C0D173C-82A1-43AB-8E9C-DE3F0067AD59}"
-	ProjectSection(ProjectDependencies) = postProject
-		{280015A3-83D7-416F-A181-7CF88BCCE12F} = {280015A3-83D7-416F-A181-7CF88BCCE12F}
-		{85DB89B5-DEA0-4429-B41E-0CAA7DB907BA} = {85DB89B5-DEA0-4429-B41E-0CAA7DB907BA}
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{280015A3-83D7-416F-A181-7CF88BCCE12F}.Debug|x64.ActiveCfg = Debug|x64
-		{280015A3-83D7-416F-A181-7CF88BCCE12F}.Debug|x64.Build.0 = Debug|x64
-		{280015A3-83D7-416F-A181-7CF88BCCE12F}.Debug|x86.ActiveCfg = Debug|Win32
-		{280015A3-83D7-416F-A181-7CF88BCCE12F}.Debug|x86.Build.0 = Debug|Win32
-		{280015A3-83D7-416F-A181-7CF88BCCE12F}.Release|x64.ActiveCfg = Release|x64
-		{280015A3-83D7-416F-A181-7CF88BCCE12F}.Release|x64.Build.0 = Release|x64
-		{280015A3-83D7-416F-A181-7CF88BCCE12F}.Release|x86.ActiveCfg = Release|Win32
-		{280015A3-83D7-416F-A181-7CF88BCCE12F}.Release|x86.Build.0 = Release|Win32
-		{85DB89B5-DEA0-4429-B41E-0CAA7DB907BA}.Debug|x64.ActiveCfg = Debug|x64
-		{85DB89B5-DEA0-4429-B41E-0CAA7DB907BA}.Debug|x64.Build.0 = Debug|x64
-		{85DB89B5-DEA0-4429-B41E-0CAA7DB907BA}.Debug|x86.ActiveCfg = Debug|Win32
-		{85DB89B5-DEA0-4429-B41E-0CAA7DB907BA}.Debug|x86.Build.0 = Debug|Win32
-		{85DB89B5-DEA0-4429-B41E-0CAA7DB907BA}.Release|x64.ActiveCfg = Release|x64
-		{85DB89B5-DEA0-4429-B41E-0CAA7DB907BA}.Release|x64.Build.0 = Release|x64
-		{85DB89B5-DEA0-4429-B41E-0CAA7DB907BA}.Release|x86.ActiveCfg = Release|Win32
-		{85DB89B5-DEA0-4429-B41E-0CAA7DB907BA}.Release|x86.Build.0 = Release|Win32
-		{3C0D173C-82A1-43AB-8E9C-DE3F0067AD59}.Debug|x64.ActiveCfg = Debug|x64
-		{3C0D173C-82A1-43AB-8E9C-DE3F0067AD59}.Debug|x64.Build.0 = Debug|x64
-		{3C0D173C-82A1-43AB-8E9C-DE3F0067AD59}.Debug|x86.ActiveCfg = Debug|Win32
-		{3C0D173C-82A1-43AB-8E9C-DE3F0067AD59}.Debug|x86.Build.0 = Debug|Win32
-		{3C0D173C-82A1-43AB-8E9C-DE3F0067AD59}.Release|x64.ActiveCfg = Release|x64
-		{3C0D173C-82A1-43AB-8E9C-DE3F0067AD59}.Release|x64.Build.0 = Release|x64
-		{3C0D173C-82A1-43AB-8E9C-DE3F0067AD59}.Release|x86.ActiveCfg = Release|Win32
-		{3C0D173C-82A1-43AB-8E9C-DE3F0067AD59}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A2B87694-5EB1-4C97-9262-7E18C4A3702E}
-	EndGlobalSection
-EndGlobal
