--- vs/game/game.hpp
@@ -1,1 +1,0 @@
-#include "32blit.hpp"
